QueryからGithubRepository[]への変換処理を共通化
